requirements.txt
rdkit-pypi==2022.9.3
streamlit==1.16.0
